sideload command tweaks
