+++ packages/office-addin-debugging/package.json
@@ -38,7 +38,7 @@
     "@types/node": "^12.12.34",
     "@types/node-fetch": "^2.5.5",
     "@types/ws": "^6.0.4",
-    "concurrently": "^3.5.0",
+    "concurrently": "^5.3.0",
     "express": "^4.16.4",
     "mocha": "^7.1.1",
     "rimraf": "^2.7.1",
