+++ packages/office-addin-dev-settings/src/main.ts
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 export * from "./appcontainer";
+export * from "./appType";
 export * from "./dev-settings";
 export * from "./process";
 export * from "./sideload";
